Reorganize entity definitions.
