+++ nuxt/components/navigation/UserMenu.vue
@@ -165,10 +165,10 @@
 import { Component, Vue } from 'nuxt-property-decorator';
 import AppChangelog from '@/components/notifications/AppChangelog.vue';
 import BugReportForm from '@/components/BugReportForm.vue';
-import { User, Role } from '@/api/auth';
+import { User, Role } from '@/entities/user';
 import LoginForm from '@/components/auth/LoginForm.vue';
 import RegisterForm from '@/components/auth/RegisterForm.vue';
-import { Feature } from '@/api/features';
+import { Feature } from '@/entities/feature';
 
 @Component({
   components: {
