+++ nuxt/entities/feature.ts
@@ -1,0 +1,8 @@
+export type FeatureMap = {
+  [name in Feature]?: boolean;
+};
+
+export enum Feature {
+  PublicUserRegistration = 'registration.public',
+  SocialAuthentication = 'auth.social',
+}
