+++ nuxt/entities/lyrics.ts
