+++ nuxt/entities/reciter.ts
@@ -1,0 +1,9 @@
+import { PersistedEntity, TimestampedEntity } from '@/entities/common';
+
+export interface Reciter extends PersistedEntity, TimestampedEntity {
+  name: string;
+  description: string | null;
+  slug: string;
+  avatar: string | null;
+  related?: { albums: number };
+}
