+++ nuxt/pages/reciters/_id.vue
@@ -104,7 +104,7 @@ import SkeletonCardGrid from '@/components/loaders/SkeletonCardGrid.vue';
 import AlbumSkeleton from '@/components/loaders/AlbumSkeleton.vue';
 import TrackCardSkeleton from '@/components/loaders/TrackCardSkeleton.vue';
 import HeroBanner from '@/components/HeroBanner.vue';
-import { Reciter } from '@/api/reciters';
+import { Reciter } from '@/entities/reciter';
 import { MetaInfo } from 'vue-meta';
 import Album from '@/components/albums/Album.vue';
 import { TrackIncludes } from '@/api/tracks';
