+++ nuxt/store/features.ts
@@ -1,5 +1,5 @@
 import { ActionContext, ActionTree, MutationTree, GetterTree } from 'vuex';
-import { Feature, FeatureMap } from '@/api/features';
+import { Feature, FeatureMap } from '@/entities/feature';
 import { RootState } from '~/store/index';
 
 /*
